v1.0.0:Datasource相关函数

+++ src/main/java/com/robby/app/commons/properties/db/DruidProperties.java
@@ -18,9 +18,9 @@ public class DruidProperties {
     int maxIdle;
     int maxActive;
     int minIdle;
-    int maxWait;
-    int timeBetweenEvictionRunsMillis;
-    int minEvictableIdleTImeMillis;
+    long maxWait;
+    long timeBetweenEvictionRunsMillis;
+    long minEvictableIdleTImeMillis;
     String validationQuery;
     Boolean testOnBorrow;
     Boolean testOnReturn;
